% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies.R
\name{anomalies}
\alias{anomalies}
\title{Title}
\usage{
anomalies(df, index = NULL, x = NULL, stl_args = list(),
  confidence = 0.99, direction = "both")
}
\arguments{
\item{df}{input data.frame}

\item{index}{optional column name of index variable. default is NULL which
uses the first column in the input data.frame}

\item{x}{optional column name of numeric variable to analyze. efault is NULL
which uses the second column in the input data.frame}

\item{stl_args}{list of arguments to pass to \link[stats]{stl} function}

\item{confidence}{anomaly threshold confidence level. higher levels increase
the anomaly threshold}

\item{direction}{direction of anomaly idendification. Options are `positive`,
`negative`, or `both`}
}
\value{
data.frame with input dataset, stl components and flag for anomalies
}
\description{
Title
}
